REFAC: RenderMaterial now uses doubles

See merge request iMSTK/iMSTK!570

+++ Source/Rendering/Materials/imstkRenderMaterial.h
@@ -110,14 +110,14 @@ public:
     ///
     /// \brief Get/Set line width or the wireframe
     ///
-    const float getLineWidth() const { return m_lineWidth; }
-    void setLineWidth(const float width);
+    const double getLineWidth() const { return m_lineWidth; }
+    void setLineWidth(const double width);
 
     ///
     /// \brief Get/Set point radius
     ///
-    const float getPointSize() const { return m_pointSize; }
-    void setPointSize(const float size);
+    const double getPointSize() const { return m_pointSize; }
+    void setPointSize(const double size);
 
     ///
     /// \brief Backface culling on/off
@@ -145,8 +145,8 @@ public:
     ///
     /// \brief Get/Set the specular amount
     ///
-    float getSpecular() const { return m_specular; }
-    void setSpecular(float specular) { m_specular = specular; }
+    double getSpecular() const { return m_specular; }
+    void setSpecular(double specular) { m_specular = specular; }
 
     ///
     /// \brief Get/Set the ambient color
@@ -157,38 +157,38 @@ public:
     ///
     /// \brief Get/Set the metalness
     ///
-    const float getMetalness() const { return m_metalness; }
-    void setMetalness(const float metalness);
+    const double getMetalness() const { return m_metalness; }
+    void setMetalness(const double metalness);
 
     ///
     /// \brief Get/Set ambient light coefficient
     ///
-    const float getAmbient() const { return m_ambient; };
-    void setAmbient(float ambient) { m_ambient = ambient; }
+    const double getAmbient() const { return m_ambient; };
+    void setAmbient(double ambient) { m_ambient = ambient; }
 
     ///
     /// \brief Get/Set diffuse light coefficient
     ///
-    const float getDiffuse() const { return m_diffuse; }
-    void setDiffuse(float diffuse) { m_diffuse = diffuse; }
+    const double getDiffuse() const { return m_diffuse; }
+    void setDiffuse(double diffuse) { m_diffuse = diffuse; }
 
     ///
     /// \brief Get/Set ambient light coefficient
     ///
-    const float getSpecularPower() const { return m_specularPower; }
-    void setSpecularPower(const float p) { m_specularPower = p; }
+    const double getSpecularPower() const { return m_specularPower; }
+    void setSpecularPower(const double p) { m_specularPower = p; }
 
     ///
     /// \brief Get/Set the roughness
     ///
-    const float getRoughness() const { return m_roughness; }
-    void setRoughness(const float roughness);
+    const double getRoughness() const { return m_roughness; }
+    void setRoughness(const double roughness);
 
     ///
     /// \brief Get/Set emissivity
     ///
-    const float getEmissivity() const { return m_emissivity; }
-    void setEmissivity(const float emissivity);
+    const double getEmissivity() const { return m_emissivity; }
+    void setEmissivity(const double emissivity);
 
     ///
     /// \brief Add/Remove/Get texture
@@ -235,14 +235,14 @@ public:
     const ShadingModel getShadingModel() const { return m_shadingModel; }
     void setShadingModel(const ShadingModel& model);
 
-    float getOcclusionStrength() const { return m_occlusionStrength; }
-    void setOcclusionStrength(const float o);
+    double getOcclusionStrength() const { return m_occlusionStrength; }
+    void setOcclusionStrength(const double o);
 
-    float getNormalStrength() const { return m_normalStrength; }
-    void setNormalStrength(const float n);
+    double getNormalStrength() const { return m_normalStrength; }
+    void setNormalStrength(const double n);
 
-    float getIndexOfRefraction() const { return m_indexOfRefraction; }
-    void setIndexOfRefraction(const float n);
+    double getIndexOfRefraction() const { return m_indexOfRefraction; }
+    void setIndexOfRefraction(const double n);
 
     const Color& getEdgeColor() const { return m_edgeColor; }
     void setEdgeColor(const Color& color);
@@ -251,7 +251,7 @@ public:
     void setVertexColor(const Color& color);
 
     const double getOpacity() const { return m_opacity; }
-    void setOpacity(const float opacity);
+    void setOpacity(const double opacity);
 
     const bool getBackfaceCulling() const { return m_backfaceCulling; }
     void setBackfaceCulling(const bool c);
@@ -287,35 +287,35 @@ protected:
     BlendMode m_blendMode = BlendMode::Alpha;
 
     ///-------------------Common properties---------------------
-    float m_diffuse      = 1.0f;
-    Color m_diffuseColor = Color::LightGray;
+    double m_diffuse      = 1.0;
+    Color  m_diffuseColor = Color::LightGray;
 
-    float m_specularPower = 100.0f;
-    float m_specular      = 0.0f;
-    Color m_specularColor = Color::Red;
+    double m_specularPower = 100.0;
+    double m_specular      = 0.0;
+    Color  m_specularColor = Color::Red;
 
-    float m_ambient      = 0.1f;
-    Color m_ambientColor = Color::White;
+    double m_ambient      = 0.1;
+    Color  m_ambientColor = Color::White;
 
-    float m_opacity = 1.0f;
+    double m_opacity = 1.0;
 
     ///-------------Wireframe specific properties----------------
-    float m_lineWidth        = 1.f;
-    float m_pointSize        = 2.f;
-    Color m_edgeColor        = Color::Marigold;
-    Color m_vertexColor      = Color::Teal;
-    bool  m_edgeVisibility   = true; ///< \note not used (vtk backend)
-    bool  m_vertexVisibility = true; ///< \note not used (vtk backend)
+    double m_lineWidth        = 1.0;
+    double m_pointSize        = 2.0;
+    Color  m_edgeColor        = Color::Marigold;
+    Color  m_vertexColor      = Color::Teal;
+    bool   m_edgeVisibility   = true; ///< \note not used (vtk backend)
+    bool   m_vertexVisibility = true; ///< \note not used (vtk backend)
 
     ///----------------PBR specific properties-------------------
-    float m_emissivity    = 0.f;
-    Color m_emmisiveColor = Color::White;
+    double m_emissivity    = 0.0;
+    Color  m_emmisiveColor = Color::White;
 
-    float m_metalness = 1.0f; ///< Value for metalness with range: [0.0, 1.0]
-    float m_roughness = 1.0f; ///< Value for roughness with range: [0.0, 1.0]
-    float m_occlusionStrength = 1.0f;
-    float m_normalStrength    = 1.0f;
-    float m_indexOfRefraction = 1.0f;
+    double m_metalness = 1.0; ///< Value for metalness with range: [0.0, 1.0]
+    double m_roughness = 1.0; ///< Value for roughness with range: [0.0, 1.0]
+    double m_occlusionStrength = 1.0;
+    double m_normalStrength    = 1.0;
+    double m_indexOfRefraction = 1.0;
 
     ///---------------------Global states------------------------
     bool m_imageBasedLighting = false;
